// src/types/anime.ts
export interface MangaCard {
  id: number;
  title: string;
  image: string;
  score: number | null;
  chapters: number | null;
  volumes: number | null;
  status: string;
  genres: string[];
  synopsis: string | null;
  authors: string[];
}